feat: Initialize core backend services with Express, Mongoose models, API routes, and Arduino compilation capabilities.
