+++ .gitignore
@@ -137,3 +137,6 @@ temp/
 *.ino
 *.hex
 test_comp.js
+
+.env
+env
