+++ package.json
@@ -2,10 +2,11 @@
   "name": "openhw-studio-backend",
   "version": "1.0.0",
   "description": "Backend for OpenHW-Studio - Electronics Simulation Platform",
-  "main": "server.js",
+  "main": "src/server.js",
+  "type": "module",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js"
   },
   "keywords": ["electronics", "simulation", "education"],
   "author": "FOSSEE",
