+++ src/models/Progress.js
@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 
 const progressSchema = new mongoose.Schema({
   userId: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true, unique: true },
@@ -8,4 +8,4 @@ const progressSchema = new mongoose.Schema({
   badges: [String]
 });
 
-module.exports = mongoose.model("Progress", progressSchema);
+export default mongoose.model("Progress", progressSchema);
